requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pyjwt
cryptography
pydantic
pydantic-settings